consolidated-ui/backend/requirements.txt
# Core FastAPI dependencies - CVE-2024-47874 mitigation
fastapi>=0.121.0
uvicorn[standard]>=0.30.0
gunicorn>=22.0.0

# Database (P2_T2: SQLAlchemy ORM + CRUD)
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.30
alembic>=1.13.0

# Redis
redis>=5.0.0
aioredis>=2.0.1

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# Validation
pydantic>=2.8.0
pydantic-settings>=2.3.0
email-validator>=2.1.0

# HTTP & async
httpx>=0.27.0
aiofiles>=23.2.0

# Rate limiting
slowapi>=0.1.9

# CORS
# (included in FastAPI)

# Utilities
python-dotenv>=1.0.0

# YAML handling (P4_T1: schedule_config.yml sync)
PyYAML>=6.0.1

# Cron scheduling
croniter>=2.0.1

# File system watcher (P4_T1: real-time YAML sync)
watchdog>=4.0.0
